refactor(node): split sync.go into focused files
Extract reorg, disconnect, and mempool snapshot logic from sync.go
(798 lines) into dedicated files for better navigability. Pure move,
no behavioral changes.

- sync.go (371 lines): config, init, accessors, core apply, rollback
- sync_reorg.go (263 lines): ApplyBlockWithReorg, branch collection
- sync_disconnect.go (124 lines): DisconnectTip, disconnect-to-ancestor
- sync_mempool.go (60 lines): mempool snapshot/restore/clone
